Fused weiss.c into uci.c
